pyproject.toml
[project]
name = "select-moe"
version = "0.1.0"
description = "..."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hydra-core",
    "torch==2.6.0",
    "transformers",
    "peft",
    "accelerate",
    "lm-eval",
    "torchvision>=0.21.0",
    "datasets",
    "scipy>=1.16.0",
    "matplotlib>=3.10.5",
    "seaborn>=0.13.2",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu124", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu124", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true



[tool.ruff]
line-length = 120

[tool.ruff.lint]
# 选择一个更保守、更安全的规则集
select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "W",  # pycodestyle warnings
    "I",  # isort (整理 imports)
    "B",  # flake8-bugbear (发现潜在bug)
    "A",  # flake8-builtins (防止覆盖内置函数)
]

[tool.ruff.lint.isort]
known-first-party = ["select-moe"]
